backend/requirements.txt
Flask==3.1.0
gunicorn==23.0.0
supabase==2.13.0
python-dotenv==1.0.1